refactor(backend): split repository sync responsibilities
