+++ backend/internal/api/readingorders_repository_sync.go
@@ -2,13 +2,9 @@ package api
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
-	"net/url"
 	"path"
 	"sort"
 	"strings"
@@ -16,8 +12,6 @@ import (
 	"time"
 
 	"github.com/Lofter1/ComicHero/backend/internal/metron"
-	"github.com/danielgtaylor/huma/v2"
-	"github.com/danielgtaylor/huma/v2/sse"
 	"github.com/jmoiron/sqlx"
 )
 
@@ -168,168 +162,6 @@ func (s *cblRepositorySyncer) Stop() {
 		s.shutdown()
 	}
 	s.stop("server stopped")
-}
-
-func defaultCBLRepositorySyncSettings() CBLRepositorySyncSettings {
-	return CBLRepositorySyncSettings{
-		Repositories: []string{"https://github.com/DieselTech/CBL-ReadingLists"},
-		Schedule:     "daily",
-		Weekdays:     []string{"monday"},
-		StartTime:    "04:00",
-	}
-}
-
-func loadCBLRepositorySyncSettings(ctx context.Context, db *sqlx.DB) (CBLRepositorySyncSettings, error) {
-	settings := defaultCBLRepositorySyncSettings()
-	var value string
-	if err := db.GetContext(ctx, &value, `SELECT value FROM app_settings WHERE key = ?`, cblRepositorySettingsKey); err != nil {
-		if err == sql.ErrNoRows {
-			return settings, nil
-		}
-		return settings, err
-	}
-	if err := json.Unmarshal([]byte(value), &settings); err != nil {
-		return settings, err
-	}
-	return settings, nil
-}
-
-func validateCBLRepositorySyncSettings(settings *CBLRepositorySyncSettings) error {
-	if settings == nil {
-		return errors.New("settings are required")
-	}
-	settings.Schedule = strings.ToLower(strings.TrimSpace(settings.Schedule))
-	if settings.Schedule != "daily" && settings.Schedule != "weekly" {
-		return errors.New("schedule must be daily or weekly")
-	}
-	if _, err := time.Parse("15:04", settings.StartTime); err != nil {
-		return errors.New("startTime must use HH:MM")
-	}
-	if len(settings.Repositories) == 0 {
-		return errors.New("at least one repository is required")
-	}
-	seenRepositories := map[string]bool{}
-	configuredRepositories := map[string]string{}
-	repositories := make([]string, 0, len(settings.Repositories))
-	for _, value := range settings.Repositories {
-		repository, err := parseCBLGitHubRepository(value)
-		if err != nil {
-			return err
-		}
-		key := strings.ToLower(repository.URL)
-		if !seenRepositories[key] {
-			seenRepositories[key] = true
-			repositories = append(repositories, repository.URL)
-		}
-		configuredRepositories[key] = repository.URL
-	}
-	settings.Repositories = repositories
-	folders := make([]CBLRepositoryFolderSelection, 0, len(settings.Folders))
-	seenFolders := map[string]bool{}
-	for _, selection := range settings.Folders {
-		repository, err := parseCBLGitHubRepository(selection.RepositoryURL)
-		if err != nil {
-			return err
-		}
-		repositoryURL, ok := configuredRepositories[strings.ToLower(repository.URL)]
-		if !ok {
-			return fmt.Errorf("folder repository %q is not configured", repository.URL)
-		}
-		folderPath, err := normalizeCBLRepositoryFolder(selection.Path)
-		if err != nil {
-			return err
-		}
-		key := strings.ToLower(repositoryURL) + "\n" + folderPath
-		if seenFolders[key] {
-			continue
-		}
-		seenFolders[key] = true
-		folders = append(folders, CBLRepositoryFolderSelection{RepositoryURL: repositoryURL, Path: folderPath})
-	}
-	sort.Slice(folders, func(i, j int) bool {
-		if !strings.EqualFold(folders[i].RepositoryURL, folders[j].RepositoryURL) {
-			return strings.ToLower(folders[i].RepositoryURL) < strings.ToLower(folders[j].RepositoryURL)
-		}
-		return strings.ToLower(folders[i].Path) < strings.ToLower(folders[j].Path)
-	})
-	settings.Folders = removeRedundantCBLRepositoryFolders(folders)
-	seenDays := map[string]bool{}
-	days := make([]string, 0, len(settings.Weekdays))
-	for _, value := range settings.Weekdays {
-		day := strings.ToLower(strings.TrimSpace(value))
-		if _, ok := weekdayNames[day]; !ok {
-			return fmt.Errorf("invalid weekday %q", day)
-		}
-		if !seenDays[day] {
-			seenDays[day] = true
-			days = append(days, day)
-		}
-	}
-	sort.Strings(days)
-	settings.Weekdays = days
-	if settings.AutoSync && settings.Schedule == "weekly" && len(days) == 0 {
-		return errors.New("weekly schedules need at least one weekday")
-	}
-	return nil
-}
-
-func normalizeCBLRepositoryFolder(value string) (string, error) {
-	value = strings.Trim(strings.TrimSpace(value), "/")
-	if value == "" || strings.Contains(value, "\\") {
-		return "", fmt.Errorf("folder path %q must be a repository-relative subfolder", value)
-	}
-	cleaned := path.Clean(value)
-	if cleaned == "." || cleaned == ".." || strings.HasPrefix(cleaned, "../") {
-		return "", fmt.Errorf("folder path %q must be a repository-relative subfolder", value)
-	}
-	return cleaned, nil
-}
-
-func removeRedundantCBLRepositoryFolders(folders []CBLRepositoryFolderSelection) []CBLRepositoryFolderSelection {
-	filtered := make([]CBLRepositoryFolderSelection, 0, len(folders))
-	for _, folder := range folders {
-		redundant := false
-		for _, selected := range filtered {
-			if strings.EqualFold(selected.RepositoryURL, folder.RepositoryURL) && pathIsInCBLRepositoryFolder(folder.Path, selected.Path) {
-				redundant = true
-				break
-			}
-		}
-		if !redundant {
-			filtered = append(filtered, folder)
-		}
-	}
-	return filtered
-}
-
-func parseCBLGitHubRepository(value string) (cblGitHubRepository, error) {
-	value = strings.TrimSpace(value)
-	parsed, err := url.Parse(value)
-	if err != nil || parsed.Scheme != "https" || !strings.EqualFold(parsed.Hostname(), "github.com") {
-		return cblGitHubRepository{}, fmt.Errorf("repository %q must be an https://github.com/owner/repository URL", value)
-	}
-	parts := strings.Split(strings.Trim(parsed.Path, "/"), "/")
-	if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-		return cblGitHubRepository{}, fmt.Errorf("repository %q must identify one GitHub repository", value)
-	}
-	name := strings.TrimSuffix(parts[1], ".git")
-	if name == "" || strings.ContainsAny(parts[0]+name, "?#") {
-		return cblGitHubRepository{}, fmt.Errorf("repository %q is invalid", value)
-	}
-	return cblGitHubRepository{
-		URL:   "https://github.com/" + parts[0] + "/" + name,
-		Owner: parts[0],
-		Name:  name,
-	}, nil
-}
-
-func saveCBLRepositorySyncSettings(ctx context.Context, db *sqlx.DB, settings CBLRepositorySyncSettings) error {
-	value, err := json.Marshal(settings)
-	if err != nil {
-		return err
-	}
-	_, err = db.ExecContext(ctx, `INSERT INTO app_settings (key, value) VALUES (?, ?) ON CONFLICT(key) DO UPDATE SET value = excluded.value`, cblRepositorySettingsKey, string(value))
-	return err
 }
 
 func (s *cblRepositorySyncer) scheduleLoop(ctx context.Context) {
@@ -749,499 +581,4 @@ func (s *cblRepositorySyncer) syncMultipartGroup(ctx context.Context, repository
 		}
 		cleanupCBLReadingOrders(ctx, s.db, []int{legacyID})
 	}
-}
-
-func (s *cblRepositorySyncer) resolveMissingComicFromMetron(ctx context.Context, book cblBook) (*Comic, error) {
-	if s.metronClient == nil {
-		return nil, errors.New("metron is not configured")
-	}
-
-	var candidates []metron.Issue
-	comicVineID, hasComicVineID := cblComicVineID(book)
-	var err error
-	if hasComicVineID {
-		candidates, err = s.metronClient.SearchIssuesByComicVineID(ctx, comicVineID)
-		if err != nil {
-			return nil, fmt.Errorf("search Metron by Comic Vine ID %d: %w", comicVineID, err)
-		}
-	}
-	if len(candidates) == 0 {
-		series := strings.TrimSpace(book.Series)
-		number := strings.TrimSpace(book.Number)
-		if series == "" || number == "" {
-			return nil, nil
-		}
-		candidates, err = s.metronClient.SearchIssues(ctx, "", series, number)
-		if err != nil {
-			return nil, fmt.Errorf("search Metron for %s #%s: %w", series, number, err)
-		}
-	}
-	candidates = uniqueMetronIssueCandidates(candidates)
-	if len(candidates) == 0 {
-		return nil, nil
-	}
-
-	selected := candidates[0]
-	if len(candidates) > 1 {
-		selected, err = s.waitForMetronIssueResolution(ctx, book, comicVineID, candidates)
-		if err != nil {
-			return nil, err
-		}
-		if selected.ID == 0 {
-			return nil, nil
-		}
-	}
-
-	detail, err := s.metronClient.GetIssue(ctx, selected.ID)
-	if err != nil {
-		return nil, fmt.Errorf("load Metron issue %d: %w", selected.ID, err)
-	}
-	output, err := importMetronComicWithOptions(ctx, s.db, s.metronClient, s.covers, *detail, MetronImportOptions{Mode: "quick"})
-	if err != nil {
-		return nil, err
-	}
-	comic := output.Body.Comic
-	return &comic, nil
-}
-
-func uniqueMetronIssueCandidates(issues []metron.Issue) []metron.Issue {
-	seen := map[int]bool{}
-	unique := make([]metron.Issue, 0, len(issues))
-	for _, issue := range issues {
-		if issue.ID <= 0 || seen[issue.ID] {
-			continue
-		}
-		seen[issue.ID] = true
-		unique = append(unique, issue)
-	}
-	return unique
-}
-
-func (s *cblRepositorySyncer) waitForMetronIssueResolution(ctx context.Context, book cblBook, comicVineID int, candidates []metron.Issue) (metron.Issue, error) {
-	byID := make(map[int]metron.Issue, len(candidates))
-	items := make([]CBLMetronIssueCandidate, 0, len(candidates))
-	for _, issue := range candidates {
-		byID[issue.ID] = issue
-		number := issue.Issue
-		if number == "" {
-			number = issue.Number
-		}
-		items = append(items, CBLMetronIssueCandidate{
-			ID:          issue.ID,
-			ComicVineID: issue.ComicVineID,
-			Series:      issue.Series,
-			SeriesYear:  issue.SeriesYear,
-			Number:      number,
-			Publisher:   issue.Publisher,
-			CoverDate:   issue.CoverDate,
-			CoverImage:  issue.CoverImage,
-		})
-	}
-
-	s.mu.Lock()
-	s.nextResolutionID++
-	resolutionID := fmt.Sprintf("cbl-metron-%d", s.nextResolutionID)
-	choices := s.resolutionChoices
-	s.status.PendingResolution = &CBLMetronIssueResolution{
-		ID:          resolutionID,
-		Series:      strings.TrimSpace(book.Series),
-		Number:      strings.TrimSpace(book.Number),
-		Volume:      strings.TrimSpace(book.Volume),
-		Year:        strings.TrimSpace(book.Year),
-		ComicVineID: comicVineID,
-		Candidates:  items,
-	}
-	s.mu.Unlock()
-	s.broadcast()
-	if choices == nil {
-		return metron.Issue{}, errors.New("metron issue selection is unavailable")
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return metron.Issue{}, ctx.Err()
-		case choice := <-choices:
-			if choice.ResolutionID != resolutionID {
-				continue
-			}
-			if choice.MetronIssueID == 0 {
-				return metron.Issue{}, nil
-			}
-			issue, ok := byID[choice.MetronIssueID]
-			if !ok {
-				return metron.Issue{}, errors.New("selected Metron issue is not a candidate")
-			}
-			return issue, nil
-		}
-	}
-}
-
-func (s *cblRepositorySyncer) resolveMetronIssue(resolutionID string, metronIssueID int) error {
-	s.mu.Lock()
-	if !s.status.Running || s.status.PendingResolution == nil || s.status.PendingResolution.ID != resolutionID {
-		s.mu.Unlock()
-		return errors.New("metron issue selection is no longer pending")
-	}
-	if metronIssueID != 0 {
-		valid := false
-		for _, candidate := range s.status.PendingResolution.Candidates {
-			valid = valid || candidate.ID == metronIssueID
-		}
-		if !valid {
-			s.mu.Unlock()
-			return errors.New("selected Metron issue is not a candidate")
-		}
-	}
-	choices := s.resolutionChoices
-	s.status.PendingResolution = nil
-	s.mu.Unlock()
-	if choices == nil {
-		return errors.New("metron issue selection is no longer pending")
-	}
-	choices <- cblMetronResolutionChoice{ResolutionID: resolutionID, MetronIssueID: metronIssueID}
-	s.broadcast()
-	return nil
-}
-
-func cblRepositoryPart(filePath string) (string, string, int, bool) {
-	stem := strings.TrimSuffix(path.Base(filePath), path.Ext(filePath))
-	parentName, partNumber, ok := cblMultipartPartName(stem)
-	if !ok {
-		return "", "", 0, false
-	}
-	directory := strings.ToLower(path.Dir(filePath))
-	keyName := strings.ToLower(strings.Join(strings.Fields(parentName), " "))
-	return directory + "/" + keyName, parentName, partNumber, true
-}
-
-func (s *cblRepositorySyncer) listRepositoryFiles(ctx context.Context, repository cblGitHubRepository) (string, []cblGitHubFile, error) {
-	var metadata struct {
-		DefaultBranch string `json:"default_branch"`
-	}
-	if err := s.getJSON(ctx, s.apiBase+"/repos/"+url.PathEscape(repository.Owner)+"/"+url.PathEscape(repository.Name), &metadata); err != nil {
-		return "", nil, fmt.Errorf("read %s metadata: %w", repository.URL, err)
-	}
-	if metadata.DefaultBranch == "" {
-		return "", nil, errors.New("repository has no default branch")
-	}
-	var tree struct {
-		Truncated bool `json:"truncated"`
-		Tree      []struct {
-			Path string `json:"path"`
-			Type string `json:"type"`
-			SHA  string `json:"sha"`
-			Size int64  `json:"size"`
-		} `json:"tree"`
-	}
-	treeURL := s.apiBase + "/repos/" + url.PathEscape(repository.Owner) + "/" + url.PathEscape(repository.Name) + "/git/trees/" + url.PathEscape(metadata.DefaultBranch) + "?recursive=1"
-	if err := s.getJSON(ctx, treeURL, &tree); err != nil {
-		return "", nil, fmt.Errorf("read %s tree: %w", repository.URL, err)
-	}
-	if tree.Truncated {
-		return "", nil, errors.New("GitHub returned a truncated repository tree")
-	}
-	files := make([]cblGitHubFile, 0)
-	for _, item := range tree.Tree {
-		if item.Type == "blob" && strings.EqualFold(path.Ext(item.Path), ".cbl") && item.Size <= cblRepositoryMaxFileSize {
-			files = append(files, cblGitHubFile{Path: item.Path, SHA: item.SHA, Size: item.Size})
-		}
-	}
-	return metadata.DefaultBranch, files, nil
-}
-
-func (s *cblRepositorySyncer) availableRepositoryFiles(ctx context.Context, settings CBLRepositorySyncSettings) ([]CBLRepositoryFile, error) {
-	available := make([]CBLRepositoryFile, 0)
-	for _, value := range settings.Repositories {
-		repository, err := parseCBLGitHubRepository(value)
-		if err != nil {
-			return nil, err
-		}
-		_, files, err := s.listRepositoryFiles(ctx, repository)
-		if err != nil {
-			return nil, err
-		}
-		groupCounts := map[string]int{}
-		for _, file := range files {
-			if groupKey, _, _, ok := cblRepositoryPart(file.Path); ok {
-				groupCounts[groupKey]++
-			}
-		}
-		for _, file := range files {
-			item := CBLRepositoryFile{RepositoryURL: repository.URL, Path: file.Path, Size: file.Size}
-			if groupKey, parentName, partNumber, ok := cblRepositoryPart(file.Path); ok && groupCounts[groupKey] > 1 {
-				item.MultipartGroup = parentName
-				item.Part = partNumber
-			}
-			available = append(available, item)
-		}
-	}
-	sort.Slice(available, func(i, j int) bool {
-		if available[i].RepositoryURL != available[j].RepositoryURL {
-			return strings.ToLower(available[i].RepositoryURL) < strings.ToLower(available[j].RepositoryURL)
-		}
-		return strings.ToLower(available[i].Path) < strings.ToLower(available[j].Path)
-	})
-	return available, nil
-}
-
-func (s *cblRepositorySyncer) getJSON(ctx context.Context, target string, output any) error {
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, target, nil)
-	if err != nil {
-		return err
-	}
-	request.Header.Set("Accept", "application/vnd.github+json")
-	request.Header.Set("User-Agent", "ComicHero-CBL-Importer")
-	response, err := s.httpClient.Do(request)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = response.Body.Close() }()
-	if response.StatusCode < 200 || response.StatusCode >= 300 {
-		return fmt.Errorf("GitHub returned %s", response.Status)
-	}
-	return json.NewDecoder(io.LimitReader(response.Body, 16<<20)).Decode(output)
-}
-
-func (s *cblRepositorySyncer) fetchCBLDocument(ctx context.Context, repository cblGitHubRepository, branch, filePath string) (cblImportDocument, error) {
-	target, err := url.Parse(s.rawBase)
-	if err != nil {
-		return cblImportDocument{}, err
-	}
-	target.Path = path.Join(target.Path, repository.Owner, repository.Name, branch, filePath)
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, target.String(), nil)
-	if err != nil {
-		return cblImportDocument{}, err
-	}
-	request.Header.Set("User-Agent", "ComicHero-CBL-Importer")
-	response, err := s.httpClient.Do(request)
-	if err != nil {
-		return cblImportDocument{}, fmt.Errorf("download %s: %w", filePath, err)
-	}
-	defer func() { _ = response.Body.Close() }()
-	if response.StatusCode < 200 || response.StatusCode >= 300 {
-		return cblImportDocument{}, fmt.Errorf("download %s: GitHub returned %s", filePath, response.Status)
-	}
-	content, err := io.ReadAll(io.LimitReader(response.Body, cblRepositoryMaxFileSize+1))
-	if err != nil {
-		return cblImportDocument{}, err
-	}
-	if len(content) > cblRepositoryMaxFileSize {
-		return cblImportDocument{}, fmt.Errorf("%s exceeds the CBL file size limit", filePath)
-	}
-	input := &ReadingOrderCBLImportInput{}
-	input.Body.Filename = filePath
-	input.Body.Content = string(content)
-	documents, err := parseCBLImportDocuments(input)
-	if err != nil {
-		return cblImportDocument{}, err
-	}
-	return documents[0], nil
-}
-
-func (s *cblRepositorySyncer) repositoryFileStates(ctx context.Context, repositoryURL string) (map[string]cblRepositoryFileState, error) {
-	rows := []cblRepositoryFileState{}
-	if err := s.db.SelectContext(ctx, &rows, `SELECT repository_url, file_path, content_sha, reading_order_id, group_key FROM cbl_repository_files WHERE repository_url = ?`, repositoryURL); err != nil {
-		return nil, err
-	}
-	states := make(map[string]cblRepositoryFileState, len(rows))
-	for _, row := range rows {
-		states[row.FilePath] = row
-	}
-	return states, nil
-}
-
-func (s *cblRepositorySyncer) saveRepositoryFileState(ctx context.Context, repositoryURL string, file cblGitHubFile, readingOrderID int, groupKey string) error {
-	_, err := s.db.ExecContext(ctx, `
-		INSERT INTO cbl_repository_files (repository_url, file_path, content_sha, reading_order_id, group_key, imported_at)
-		VALUES (?, ?, ?, ?, ?, ?)
-		ON CONFLICT(repository_url, file_path) DO UPDATE SET
-			content_sha = excluded.content_sha,
-			reading_order_id = excluded.reading_order_id,
-			group_key = excluded.group_key,
-			imported_at = excluded.imported_at
-	`, repositoryURL, file.Path, file.SHA, readingOrderID, groupKey, currentTimestamp())
-	return err
-}
-
-func (s *cblRepositorySyncer) setCurrentFile(file string) {
-	s.update(func(status *CBLRepositorySyncStatus) { status.CurrentFile = file })
-}
-
-func (s *cblRepositorySyncer) increment(kind string) {
-	s.update(func(status *CBLRepositorySyncStatus) {
-		switch kind {
-		case "imported":
-			status.Imported++
-		case "updated":
-			status.Updated++
-		default:
-			status.Unchanged++
-		}
-	})
-}
-
-func (s *cblRepositorySyncer) recordFailure(err error) {
-	if err == nil {
-		return
-	}
-	s.update(func(status *CBLRepositorySyncStatus) {
-		status.Failed++
-		status.LastError = err.Error()
-	})
-}
-
-func (s *cblRepositorySyncer) update(change func(*CBLRepositorySyncStatus)) {
-	s.mu.Lock()
-	change(&s.status)
-	s.mu.Unlock()
-	s.broadcast()
-}
-
-func (s *cblRepositorySyncer) snapshot(ctx context.Context) CBLRepositorySyncStatus {
-	settings, _ := loadCBLRepositorySyncSettings(ctx, s.db)
-	s.mu.Lock()
-	status := s.status
-	s.mu.Unlock()
-	status.Settings = settings
-	return status
-}
-
-func (s *cblRepositorySyncer) subscribe(ctx context.Context) (<-chan CBLRepositorySyncStatus, func()) {
-	s.mu.Lock()
-	s.nextSubscriberID++
-	id := s.nextSubscriberID
-	updates := make(chan CBLRepositorySyncStatus, 16)
-	s.subscribers[id] = updates
-	s.mu.Unlock()
-	updates <- s.snapshot(ctx)
-	return updates, func() {
-		s.mu.Lock()
-		if current, ok := s.subscribers[id]; ok {
-			delete(s.subscribers, id)
-			close(current)
-		}
-		s.mu.Unlock()
-	}
-}
-
-func (s *cblRepositorySyncer) broadcast() {
-	status := s.snapshot(context.Background())
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	for _, updates := range s.subscribers {
-		select {
-		case updates <- status:
-		default:
-			select {
-			case <-updates:
-			default:
-			}
-			select {
-			case updates <- status:
-			default:
-			}
-		}
-	}
-}
-
-type CBLRepositorySyncStatusOutput struct{ Body CBLRepositorySyncStatus }
-type UpdateCBLRepositorySyncSettingsInput struct{ Body CBLRepositorySyncSettings }
-type CBLRepositoryFilesOutput struct{ Body []CBLRepositoryFile }
-type TriggerCBLRepositorySyncInput struct {
-	Body struct {
-		Files                  []CBLRepositoryFileSelection `json:"files,omitempty"`
-		ResolveMissingOnMetron bool                         `json:"resolveMissingOnMetron,omitempty"`
-	}
-}
-type ResolveCBLMetronIssueInput struct {
-	Body struct {
-		ResolutionID  string `json:"resolutionId" minLength:"1"`
-		MetronIssueID int    `json:"metronIssueId" minimum:"0"`
-	}
-}
-
-func RegisterCBLRepositorySyncRoutes(api huma.API, db *sqlx.DB, syncer *cblRepositorySyncer) {
-	huma.Register(api, huma.Operation{OperationID: "getCBLRepositorySync", Tags: []string{tagReadingOrders}, Summary: "Get CBL repository sync settings and status", Method: http.MethodGet, Path: "/readingOrders/repository-sync", Errors: []int{401, 403, 500}}, func(ctx context.Context, _ *struct{}) (*CBLRepositorySyncStatusOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		return &CBLRepositorySyncStatusOutput{Body: syncer.snapshot(ctx)}, nil
-	})
-	sse.Register(api, huma.Operation{OperationID: "streamCBLRepositorySync", Tags: []string{tagReadingOrders}, Summary: "Stream CBL repository sync status", Method: http.MethodGet, Path: "/readingOrders/repository-sync/events", Errors: []int{401, 403, 500}}, map[string]any{"sync": CBLRepositorySyncEvent{}}, func(ctx context.Context, _ *struct{}, send sse.Sender) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return
-		}
-		updates, unsubscribe := syncer.subscribe(ctx)
-		defer unsubscribe()
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case status, ok := <-updates:
-				if !ok || send.Data(CBLRepositorySyncEvent{Sync: status}) != nil {
-					return
-				}
-			}
-		}
-	})
-	huma.Register(api, huma.Operation{OperationID: "updateCBLRepositorySync", Tags: []string{tagReadingOrders}, Summary: "Update CBL repository sync settings", Method: http.MethodPut, Path: "/readingOrders/repository-sync", Errors: []int{400, 401, 403, 500}}, func(ctx context.Context, input *UpdateCBLRepositorySyncSettingsInput) (*CBLRepositorySyncStatusOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		if err := validateCBLRepositorySyncSettings(&input.Body); err != nil {
-			return nil, huma.Error400BadRequest(err.Error())
-		}
-		if err := saveCBLRepositorySyncSettings(ctx, db, input.Body); err != nil {
-			return nil, huma.Error500InternalServerError("failed to save CBL repository settings")
-		}
-		select {
-		case syncer.wake <- struct{}{}:
-		default:
-		}
-		syncer.broadcast()
-		return &CBLRepositorySyncStatusOutput{Body: syncer.snapshot(ctx)}, nil
-	})
-	huma.Register(api, huma.Operation{OperationID: "listCBLRepositoryFiles", Tags: []string{tagReadingOrders}, Summary: "List available repository CBL files", Method: http.MethodGet, Path: "/readingOrders/repository-sync/files", Errors: []int{401, 403, 500}}, func(ctx context.Context, _ *struct{}) (*CBLRepositoryFilesOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		settings, err := loadCBLRepositorySyncSettings(ctx, db)
-		if err != nil {
-			return nil, huma.Error500InternalServerError("failed to load CBL repository settings")
-		}
-		files, err := syncer.availableRepositoryFiles(ctx, settings)
-		if err != nil {
-			return nil, huma.Error500InternalServerError(err.Error())
-		}
-		return &CBLRepositoryFilesOutput{Body: files}, nil
-	})
-	huma.Register(api, huma.Operation{OperationID: "triggerCBLRepositorySync", Tags: []string{tagReadingOrders}, Summary: "Import all or selected CBL repository files", Method: http.MethodPost, Path: "/readingOrders/repository-sync/trigger", DefaultStatus: http.StatusAccepted, Errors: []int{400, 401, 403, 409, 500}}, func(ctx context.Context, input *TriggerCBLRepositorySyncInput) (*CBLRepositorySyncStatusOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		if err := syncer.trigger("manual", input.Body.Files, input.Body.ResolveMissingOnMetron); err != nil {
-			if strings.Contains(err.Error(), "already running") {
-				return nil, huma.Error409Conflict(err.Error())
-			}
-			return nil, huma.Error400BadRequest(err.Error())
-		}
-		return &CBLRepositorySyncStatusOutput{Body: syncer.snapshot(ctx)}, nil
-	})
-	huma.Register(api, huma.Operation{OperationID: "resolveCBLRepositoryMetronIssue", Tags: []string{tagReadingOrders}, Summary: "Choose a Metron issue for an ambiguous CBL book", Method: http.MethodPost, Path: "/readingOrders/repository-sync/resolve", Errors: []int{400, 401, 403, 409, 500}}, func(ctx context.Context, input *ResolveCBLMetronIssueInput) (*CBLRepositorySyncStatusOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		if err := syncer.resolveMetronIssue(strings.TrimSpace(input.Body.ResolutionID), input.Body.MetronIssueID); err != nil {
-			return nil, huma.Error409Conflict(err.Error())
-		}
-		return &CBLRepositorySyncStatusOutput{Body: syncer.snapshot(ctx)}, nil
-	})
-	huma.Register(api, huma.Operation{OperationID: "stopCBLRepositorySync", Tags: []string{tagReadingOrders}, Summary: "Stop CBL repository import", Method: http.MethodPost, Path: "/readingOrders/repository-sync/stop", Errors: []int{401, 403, 500}}, func(ctx context.Context, _ *struct{}) (*CBLRepositorySyncStatusOutput, error) {
-		if _, err := requireAdminUser(ctx, db); err != nil {
-			return nil, err
-		}
-		syncer.stop("stopped by admin")
-		return &CBLRepositorySyncStatusOutput{Body: syncer.snapshot(ctx)}, nil
-	})
 }
